requirements.txt
discord.py==2.3.0
dotenv-python==0.0.1
mysql-connector-python==8.0.33
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.3
python-dotenv==1.0.0
requests==2.31.0